kf_prefect/requirements.txt
boto3
boto3-stubs[full]
mypy
pyyaml
types-PyYAML
pydantic
